+++ src/ClientContextInterface.php
@@ -42,6 +42,7 @@ use Predis\Command\CommandInterface;
 use Predis\Command\Container\ACL;
 use Predis\Command\Container\CLIENT;
 use Predis\Command\Container\FUNCTIONS;
+use Predis\Command\Container\HOTKEYS;
 use Predis\Command\Container\Json\JSONDEBUG;
 use Predis\Command\Container\Search\FTCONFIG;
 use Predis\Command\Container\Search\FTCURSOR;
@@ -272,7 +273,7 @@ use Predis\Command\Redis\VADD;
  * @method $this topklist(string $key, bool $withCount = false)
  * @method $this topkquery(string $key, ...$items)
  * @method $this topkreserve(string $key, int $topK, int $width = 8, int $depth = 7, float $decay = 0.9)
- * @method $this tsadd(string $key, int $timestamp, float $value, ?AddArguments $arguments = null)
+ * @method $this tsadd(string $key, int $timestamp, string|float $value, ?AddArguments $arguments = null)
  * @method $this tsalter(string $key, ?TSAlterArguments $arguments = null)
  * @method $this tscreate(string $key, ?TSCreateArguments $arguments = null)
  * @method $this tscreaterule(string $sourceKey, string $destKey, string $aggregator, int $bucketDuration, int $alignTimestamp = 0)
@@ -396,6 +397,7 @@ use Predis\Command\Redis\VADD;
  *
  * Container commands
  * @property CLIENT    $client
+ * @property HOTKEYS   $hotkeys
  * @property FUNCTIONS $function
  * @property FTCONFIG  $ftconfig
  * @property FTCURSOR  $ftcursor